Allow customisation of the sub header icon links on EntityAboutCard
