+++ plugins/catalog/src/components/AboutCard/AboutCard.tsx
@@ -16,10 +16,7 @@
 import {
   ANNOTATION_EDIT_URL,
   ANNOTATION_LOCATION,
-  CompoundEntityRef,
-  DEFAULT_NAMESPACE,
   stringifyEntityRef,
-  parseEntityRef,
 } from '@backstage/catalog-model';
 import Card from '@material-ui/core/Card';
 import CardContent from '@material-ui/core/CardContent';
@@ -36,40 +33,24 @@ import {
 } from '@backstage/core-components';
 import React, { useCallback } from 'react';
 import {
-  ScmIntegrationIcon,
-  scmIntegrationsApiRef,
-} from '@backstage/integration-react';
-import {
   alertApiRef,
   errorApiRef,
   useApi,
-  useApp,
   useRouteRef,
 } from '@backstage/core-plugin-api';
-import {
-  catalogApiRef,
-  getEntitySourceLocation,
-  useEntity,
-} from '@backstage/plugin-catalog-react';
-import { createFromTemplateRouteRef, viewTechDocRouteRef } from '../../routes';
-
+import { catalogApiRef, useEntity } from '@backstage/plugin-catalog-react';
+import { createFromTemplateRouteRef } from '../../routes';
 import { AboutContent } from './AboutContent';
 import CachedIcon from '@material-ui/icons/Cached';
-import CreateComponentIcon from '@material-ui/icons/AddCircleOutline';
-import DocsIcon from '@material-ui/icons/Description';
 import EditIcon from '@material-ui/icons/Edit';
-import { isTemplateEntityV1beta3 } from '@backstage/plugin-scaffolder-common';
 import { useEntityPermission } from '@backstage/plugin-catalog-react/alpha';
 import { catalogEntityRefreshPermission } from '@backstage/plugin-catalog-common/alpha';
-import { useSourceTemplateCompoundEntityRef } from './hooks';
-import { taskCreatePermission } from '@backstage/plugin-scaffolder-common/alpha';
-import { usePermission } from '@backstage/plugin-permission-react';
+import {
+  useDefaultIconLinks,
+  useSourceTemplateCompoundEntityRef,
+} from './hooks';
 import { catalogTranslationRef } from '../../alpha/translation';
 import { useTranslationRef } from '@backstage/core-plugin-api/alpha';
-
-const TECHDOCS_ANNOTATION = 'backstage.io/techdocs-ref';
-
-const TECHDOCS_EXTERNAL_ANNOTATION = 'backstage.io/techdocs-entity';
 
 const useStyles = makeStyles({
   gridItemCard: {
@@ -98,6 +79,11 @@ const useStyles = makeStyles({
  */
 export interface AboutCardProps {
   variant?: InfoCardVariants;
+  /**
+   * Optional list of Icon links to display in the card subheader
+   * Providing any icons here overrides the default icons.
+   */
+  iconLinks?: IconLinkVerticalProps[];
 }
 
 /**
@@ -108,15 +94,12 @@ export interface AboutCardProps {
  * card in your own repository instead, that is perfect for your own needs.
  */
 export function AboutCard(props: AboutCardProps) {
-  const { variant } = props;
-  const app = useApp();
+  const { variant, iconLinks } = props;
   const classes = useStyles();
   const { entity } = useEntity();
-  const scmIntegrationsApi = useApi(scmIntegrationsApiRef);
   const catalogApi = useApi(catalogApiRef);
   const alertApi = useApi(alertApiRef);
   const errorApi = useApi(errorApiRef);
-  const viewTechdocLink = useRouteRef(viewTechDocRouteRef);
   const templateRoute = useRouteRef(createFromTemplateRouteRef);
   const sourceTemplateRef = useSourceTemplateCompoundEntityRef(entity);
   const { allowed: canRefresh } = useEntityPermission(
@@ -124,78 +107,10 @@ export function AboutCard(props: AboutCardProps) {
   );
   const { t } = useTranslationRef(catalogTranslationRef);
 
-  const { allowed: canCreateTemplateTask } = usePermission({
-    permission: taskCreatePermission,
-  });
+  const subHeaderLinks = useDefaultIconLinks(entity, iconLinks);
 
-  const entitySourceLocation = getEntitySourceLocation(
-    entity,
-    scmIntegrationsApi,
-  );
   const entityMetadataEditUrl =
     entity.metadata.annotations?.[ANNOTATION_EDIT_URL];
-
-  let techdocsRef: CompoundEntityRef | undefined;
-
-  if (entity.metadata.annotations?.[TECHDOCS_EXTERNAL_ANNOTATION]) {
-    try {
-      techdocsRef = parseEntityRef(
-        entity.metadata.annotations?.[TECHDOCS_EXTERNAL_ANNOTATION],
-      );
-      // not a fan of this but we don't care if the parseEntityRef fails
-    } catch {
-      techdocsRef = undefined;
-    }
-  }
-
-  const viewInSource: IconLinkVerticalProps = {
-    label: t('aboutCard.viewSource'),
-    disabled: !entitySourceLocation,
-    icon: <ScmIntegrationIcon type={entitySourceLocation?.integrationType} />,
-    href: entitySourceLocation?.locationTargetUrl,
-  };
-  const viewInTechDocs: IconLinkVerticalProps = {
-    label: t('aboutCard.viewTechdocs'),
-    disabled:
-      !(
-        entity.metadata.annotations?.[TECHDOCS_ANNOTATION] ||
-        entity.metadata.annotations?.[TECHDOCS_EXTERNAL_ANNOTATION]
-      ) || !viewTechdocLink,
-    icon: <DocsIcon />,
-    href:
-      viewTechdocLink &&
-      (techdocsRef
-        ? viewTechdocLink({
-            namespace: techdocsRef.namespace || DEFAULT_NAMESPACE,
-            kind: techdocsRef.kind,
-            name: techdocsRef.name,
-          })
-        : viewTechdocLink({
-            namespace: entity.metadata.namespace || DEFAULT_NAMESPACE,
-            kind: entity.kind,
-            name: entity.metadata.name,
-          })),
-  };
-
-  const subHeaderLinks = [viewInSource, viewInTechDocs];
-
-  if (isTemplateEntityV1beta3(entity)) {
-    const Icon = app.getSystemIcon('scaffolder') ?? CreateComponentIcon;
-
-    const launchTemplate: IconLinkVerticalProps = {
-      label: t('aboutCard.launchTemplate'),
-      icon: <Icon />,
-      disabled: !templateRoute || !canCreateTemplateTask,
-      href:
-        templateRoute &&
-        templateRoute({
-          templateName: entity.metadata.name,
-          namespace: entity.metadata.namespace || DEFAULT_NAMESPACE,
-        }),
-    };
-
-    subHeaderLinks.push(launchTemplate);
-  }
 
   let cardClass = '';
   if (variant === 'gridItem') {
