+++ plugins/catalog/src/components/AboutCard/index.ts
@@ -20,3 +20,4 @@ export { AboutContent } from './AboutContent';
 export type { AboutContentProps } from './AboutContent';
 export { AboutField } from './AboutField';
 export type { AboutFieldProps } from './AboutField';
+export { useDefaultIconLinks } from './hooks';
